+++ backends/hip-ref/ceed-hip-ref.h
@@ -109,7 +109,7 @@ typedef struct {
 typedef struct {
   hipModule_t module;
   hipFunction_t linearAssemble;
-  CeedInt nelem, nnodes, elemsPerBlock;
+  CeedInt nelem, block_size_x, block_size_y, elemsPerBlock;
   CeedScalar *d_B_in, *d_B_out;
 } CeedOperatorAssemble_Hip;